refactor: 도메인별 폴더 구조 정리
