+++ src/main/kotlin/com/wq/auth/api/controller/email/request/EmailRequestDto.kt
@@ -1,0 +1,3 @@
+package com.wq.auth.api.controller.email.request
+
+data class EmailRequestDto (val email: String)
