+++ src/main/kotlin/com/wq/auth/api/domain/email/error/EmailExceptionCode.kt
@@ -1,4 +1,4 @@
-package com.wq.auth.email.error
+package com.wq.auth.api.domain.email.error
 
 import com.wq.auth.shared.error.ApiResponseCode
