+++ src/main/kotlin/com/wq/auth/api/domain/member/MemberRepository.kt
@@ -1,6 +1,6 @@
-package com.wq.auth.repository
+package com.wq.auth.api.domain.member
 
-import com.wq.auth.entity.MemberEntity
+import com.wq.auth.api.domain.member.entity.MemberEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
