+++ src/main/kotlin/com/wq/auth/api/domain/member/entity/ProviderType.kt
@@ -1,4 +1,4 @@
-package com.wq.auth.entity
+package com.wq.auth.api.domain.member.entity
 
 enum class ProviderType {
     EMAIL,      // 자체 회원가입
