+++ src/main/kotlin/com/wq/auth/api/domain/member/entity/Role.kt
@@ -1,0 +1,6 @@
+package com.wq.auth.api.domain.member.entity
+
+enum class Role {
+    MEMBER,
+    ADMIN
+}
