--- src/main/kotlin/com/wq/auth/email/EmailRequestDto.kt
@@ -1,3 +1,0 @@
-package com.wq.auth.email
-
-data class EmailRequestDto (val email: String)
