--- src/main/kotlin/com/wq/auth/entity/Role.kt
@@ -1,6 +1,0 @@
-package com.wq.auth.entity
-
-enum class Role {
-    MEMBER,
-    ADMIN
-}
